+++ changelog.md
@@ -36,6 +36,7 @@ These changes will (most likely) be included in the next version.
 - Block explosion events cancelled by other plugins now remain cancelled unless MobArena specifically uncancels them for an arena.
 - Flaming arrows now ignite TNT blocks in the arena.
 - Players no longer take fall damage when they leave (or get removed from) an arena while falling.
+- Players no longer take damage from projectiles shot by pets of other players.
 
 ## [0.106] - 2021-05-09
 ### Added